feat: added custom header logic
